feat(slider): component and render.

+++ RGB-Color-Picker/ColorSlider.swift
@@ -10,11 +10,11 @@ import SwiftData
 
 @Model
 final class ColorSlider {
-    var red: Int
-    var green: Int
-    var blue: Int
+    var red: Double
+    var green: Double
+    var blue: Double
     
-    init(red: Int, green: Int, blue: Int) {
+    init(red: Double, green: Double, blue: Double) {
         self.red = red
         self.green = green
         self.blue = blue
